pyproject.toml
[tool.poetry]
name = "Discord Voice 24/7"
version = "0.1.0"
description = "Một Chương Trình Làm Cho Tài Khoản Discord Của Bạn Trong Phòng Voice 24/7"
authors = ["SealedSaucer"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
PyNaCl = "^1.5.0"
Flask = "^2.2.1"
"discord.py-self" = {git = "https://github.com/dolfies/discord.py-self"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
